Move ForExplorerFilter into Admin API implementation
Not generally useful in the Public API
